Install OpenSSL development headers
As it's mentioned in uWSGI docs[1]:
"In order to use https option be sure that you have OpenSSL development
headers installed (e.g. libssl-dev on Debian). Install them and rebuild
uWSGI so the build system will automatically detect it."

This change ensures that required package is installed on the host where
uWSGI is being built.

We don't need to force reinstall during an openstack upgrade because
uWSGI version was bumped in Zed[2] so uWSGI is going to be
reinstalled/upgraded anyway.

According to my tests, OpenSSL development headers are not required for
distro installs.

[1] https://uwsgi-docs.readthedocs.io/en/latest/HTTPS.html
[2] https://review.opendev.org/c/openstack/openstack-ansible/+/864579

Change-Id: Ib3cf675fa22a61ddf800af6276071410163bcf68