Enable uwsgi to bind to multiple different IP addresses
Previously the template could only insert one http-socket or
https-socket line into the uwsgi configuration file. This meant
a choice between binding to a specific IP address, or 0.0.0.0 to
listen on all addresses.

This patch makes the template distinguish between uwsgi_bind_address
being an iterable list of addresses, or a single address. It is now
possible to specify a number of specific addresses to bind to.

Change-Id: I088a64c65a9f9912091bd10402028db574767ca4